+++ news/utils.py
@@ -3,7 +3,5 @@ import os
 
 def delete_images(image_path):
     full_path = os.path.join(settings.MEDIA_ROOT, image_path)
-    try:
+    if os.path.exists(full_path):
         os.remove(full_path)
-    except FileNotFoundError:
-        print(f"El archivo {full_path} no existe.")
